Added increase and decrese cart item!

+++ src/main/java/com/deskbudd/Repo/ItemRepo.java
@@ -4,6 +4,6 @@ import com.deskbudd.models.ItemModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 @org.springframework.stereotype.Repository
-public interface ItemRepo extends JpaRepository<ItemModel, Integer> {
+public interface ItemRepo extends JpaRepository<ItemModel, Long> {
 
 }
